requirements.txt
vanna>=0.4.3
chromadb>=0.4.22
snowflake-connector-python>=3.0.0
cryptography>=41.0.0
openai>=1.1.0
flask>=2.3.0
plotly>=5.18.0 
mem0ai>=0.1.67
httpx>=0.28.1
python-dotenv>=1.0.1
streamlit>=1.32.0 